rename : [folder] selector > selects
